requirements.txt
apache-airflow==2.7.1
psycopg2-binary==2.9.6
SQLAlchemy==1.4.49
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3